+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to [ChaiLove](https://github.com/RobLoach/ChaiLove) will be 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 0.12.0 - Unreleased
+### Added
+- `SoundData.getVolume()`
+- `SoundData.setVolume()`
+
 ## 0.11.1 - 2017-12-10
 ### Fixed
 - Version number